requirements.txt
# Core web framework
fastapi==0.116.1
uvicorn==0.35.0

# AWS Bedrock integration
boto3==1.39.14
botocore==1.39.14

# JWT authentication
python-jose[cryptography]==3.5.0
requests==2.32.4

# Core dependencies (likely installed with FastAPI)
pydantic==2.10.4
starlette==0.41.3